correct splitting of test cases
