+++ src/eduid_common/api/testing.py
@@ -38,6 +38,7 @@ import logging
 import traceback
 from contextlib import contextmanager
 from copy import deepcopy
+from typing import Dict, Any
 
 from flask.testing import FlaskClient
 
@@ -80,17 +81,51 @@ TEST_CONFIG = {
 }
 
 
+_standard_test_users = {
+    'hubba-bubba': NEW_USER_EXAMPLE,
+    'hubba-baar': NEW_UNVERIFIED_USER_EXAMPLE,
+    'hubba-fooo': NEW_COMPLETED_SIGNUP_USER_EXAMPLE,
+}
+
+
+class APIMockedUserDB(object):
+
+    test_users: Dict[str, Any] = {}
+
+    def __init__(self, _patches):
+        pass
+
+    def all_userdocs(self):
+        for user in self.test_users.values():
+            yield deepcopy(user)
+
+
 class EduidAPITestCase(CommonTestCase):
     """
     Base Test case for eduID APIs.
 
     See the `load_app` and `update_config` methods below before subclassing.
     """
+    # This concept with a class variable is broken - doesn't provide isolation between tests.
+    # Do what we can and initialise it empty here, and then fill it in __init__.
+    MockedUserDB = APIMockedUserDB
 
     def setUp(self, users=None, am_settings=None):
         '''
         set up tests
         '''
+        # test users
+        self.MockedUserDB.test_users = {}
+        if users is None:
+            users = ['hubba-bubba']
+        for this in users:
+            self.MockedUserDB.test_users[this] = _standard_test_users.get(this)
+
+        self.user = None
+        # Initialize some convenience variables on self based on the first user in `users'
+        self.test_user_data = _standard_test_users.get(users[0])
+        self.test_user = User(data=self.test_user_data)
+
         super(EduidAPITestCase, self).setUp(users=users,
                                             am_settings=am_settings)
         # Set up Redis for shared sessions
